refactor: change filename to kebab-case
